+++ src/js/helpers/event-bus.js
@@ -58,6 +58,10 @@ export class EventBus extends Vue {
       error: 'error',
       info: 'info',
       updatePollRequests: 'polls:updateRequestsList',
+      customersUpdateList: 'customers:updateList',
+      customersMassIssue: 'customers:massIssue',
+      // TODO: should not be here
+      businessesSetCurrentBusiness: 'businesses:setCurrentBusiness',
     }
   }