+++ src/vue/modules/requests/poll-requests/components/request-attributes-viewer.vue
@@ -25,27 +25,6 @@
         </tr>
 
         <tr>
-          <td>{{ 'poll-requests.result-provider' | globalize }}</td>
-          <td>
-            <email-getter :account-id="request.resultProvider" />
-          </td>
-        </tr>
-
-        <tr>
-          <!-- eslint-disable-next-line max-len -->
-          <td>{{ 'poll-requests.vote-confirmation-required-title' | globalize }}</td>
-          <td>
-            <template v-if="request.voteConfirmationRequired">
-              {{ 'poll-requests.yes-msg' | globalize }}
-            </template>
-
-            <template v-else>
-              {{ 'poll-requests.no-msg' | globalize }}
-            </template>
-          </td>
-        </tr>
-
-        <tr>
           <td
             class="request-attributes-viewer__question"
             :colspan="isQuestionMaxLen ? 2 : 1"
@@ -78,15 +57,11 @@
 import { PollRequest } from '../wrappers/poll-request'
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
-import EmailGetter from '@/vue/common/EmailGetter'
 
 const QUESTION_MAX_LEN = 20
 
 export default {
   name: 'request-attributes-viewer',
-  components: {
-    EmailGetter,
-  },
   props: {
     request: { type: PollRequest, required: true },
   },