+++ src/vue/pages/sales/SalesList.vue
@@ -172,6 +172,10 @@ export default {
           break
       }
 
+      if (this.$route.query.owner) {
+        opts.filter.owner = this.$route.query.owner
+      }
+
       return function () {
         return api.getWithSignature(endpoint, opts)
       }
@@ -181,7 +185,6 @@ export default {
   watch: {
     'recordsLoader': function () {
       this.saleRecords = []
-      this.isLoaded = false
     },
     isUserSales () {
       this.recordsLoader()