+++ CHANGELOG.md
@@ -8,13 +8,28 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+#### Fixed
+- Resolved issue when select was without margin on the changing limits form
+- Resolved issue when you can set amount much bigger than the max amount of asset
+- Resolved an issue when limit left amount not updated on the client side
+- Resolved an issue when poll don't load for a user
+- Resolved an issue when participant email doesn't show on `Participants` tab
+  on `Poll` page
+- Resolved an issue when user can't update sale request
+
+## [1.10.0-x.0] - 2019-07-02
 #### Added
 - New Atomic swap feature:
   - Create atomic swap form
 - `canBeBaseInAtomicSwap` and `canBeQuoteInAtomicSwap` policy for assets
 
-#### Fixed
-- Resolved an issue when user can't update sale request
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.9.0-x.0
+
+### Experimental features changes
+#### Added
+- New use case - conto
 
 ## [1.9.0-rc.0] - 2019-06-24
 #### Added
@@ -810,7 +825,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.9.0-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.10.0-x.0...HEAD
+[1.10.0-x.0]: https://github.com/tokend/web-client/compare/1.9.0-rc.0...1.10.0-x.0
 [1.9.0-rc.0]: https://github.com/tokend/web-client/compare/1.8.0...1.9.0-rc.0
 [1.8.0]: https://github.com/tokend/web-client/compare/1.8.0-rc.3...1.8.0
 [1.8.0-rc.3]: https://github.com/tokend/web-client/compare/1.8.0-rc.2...1.8.0-rc.3